remove old main suite, rename optimized main suite to be the new main suite
